Moved fetchRecipes to DataApi.js
