+++ src/components/Recipes.js
@@ -9,16 +9,7 @@ import {
   ItemTitle,
   Button,
 } from "./Styles.js";
-
-const fetchRecipes = () =>
-  fetch(`${process.env.REACT_APP_API_BASEURL}/api/recipe/recipes/`).then(
-    (res) => {
-      if (res.status === 200) {
-        return res.json();
-      }
-      return null;
-    }
-  );
+import api from "../data/DataAPI.js";
 
 function RecipeList({ recipes }) {
   let match = useRouteMatch();
@@ -58,7 +49,8 @@ function Recipes(props) {
 
   useEffect(() => {
     dispatch((state) => ({ ...state, isLoading: true }));
-    fetchRecipes()
+    api
+      .fetchRecipes()
       .then((data) => dispatch({ data, error: null, isLoading: false }))
       .catch((error) => dispatch({ data: null, error, isLoading: false }));
   }, []);
