+++ src/data/DataAPI.js
@@ -16,6 +16,16 @@ const postData = async (url = "", method, data = {}) => {
   return response.json(); // parses JSON response into native JavaScript objects
 };
 
+const fetchRecipes = () =>
+  fetch(`${process.env.REACT_APP_API_BASEURL}/api/recipe/recipes/`).then(
+    (res) => {
+      if (res.status === 200) {
+        return res.json();
+      }
+      return null;
+    }
+  );
+
 const fetchRecipe = (recipeId) => {
   return fetch(
     `${process.env.REACT_APP_API_BASEURL}/api/recipe/recipes/${recipeId}/`
@@ -37,6 +47,7 @@ const api = {
   postData,
   fetchRecipe,
   deleteRecipe,
+  fetchRecipes,
 };
 
 export default api;
